+++ qatrack/reports/models.py
@@ -184,7 +184,7 @@ class ReportSchedule(models.Model):
         recipients = []
         for fn, ln, e in users:
 
-            e = e.strip()
+            e = e.strip() if e else None
 
             if fn and ln and e:
                 recipients.append('"%s %s"<%s>' % (fn, ln, e))